Add variant display names to hash algorithm DynamicData tests
Wires DynamicDataDisplayName onto every [DynamicData(nameof(HashAlgorithmVariants))]
and [DynamicData(nameof(NonCryptographicHashAlgorithmVariants))] call site so
data-driven failures name the variant ("Variant: Blake2b384") instead of an
opaque row index.

Resolution is hosted on non-generic helper classes (HashAlgorithmVariantDisplayName
and NonCryptographicHashAlgorithmVariantDisplayName) and wired via
DynamicDataDisplayNameDeclaringType because MSTest 4.x resolves the display
name method through Type.GetDeclaredMethod, which does not walk inheritance --
hosting the helper on a non-generic static class lets every base, intermediate,
and concrete test class share a single implementation.

Matches the existing SymmetricAlgorithmTestDataSource / KatDisplayName pattern.
