+++ Bodu.IO.Hashing/test/IO.Hashing/FnvTests.AlgorithmName.cs
@@ -15,7 +15,7 @@ public abstract partial class FnvTests<TTest, TAlgorithm>
     /// </summary>
     /// <param name="variant">The algorithm variant under test.</param>
     [TestMethod]
-    [DynamicData(nameof(NonCryptographicHashAlgorithmVariants))]
+    [DynamicData(nameof(NonCryptographicHashAlgorithmVariants), DynamicDataDisplayName = nameof(NonCryptographicHashAlgorithmVariantDisplayName.GetDisplayName), DynamicDataDisplayNameDeclaringType = typeof(NonCryptographicHashAlgorithmVariantDisplayName))]
     public void AlgorithmName_WhenUsingVariant_ShouldReturnCorrectlyFormattedString(SingleTestVariant variant)
     {
         TAlgorithm algorithm = CreateAlgorithm(variant);
