+++ Bodu.IO.Hashing/test/IO.Hashing/NonCryptographicHashAlgorithmTests.GetHashCode.cs
@@ -18,7 +18,7 @@ public abstract partial class NonCryptographicHashAlgorithmTests<TTest, TAlgorit
     /// </summary>
     /// <param name="variant">The algorithm variant under test.</param>
     [TestMethod]
-    [DynamicData(nameof(NonCryptographicHashAlgorithmVariants))]
+    [DynamicData(nameof(NonCryptographicHashAlgorithmVariants), DynamicDataDisplayName = nameof(NonCryptographicHashAlgorithmVariantDisplayName.GetDisplayName), DynamicDataDisplayNameDeclaringType = typeof(NonCryptographicHashAlgorithmVariantDisplayName))]
     public void GetHashCode_WhenCalled_ShouldThrowExactly(TVariant variant)
     {
         TAlgorithm algorithm = CreateAlgorithm(variant);
