+++ Bodu.Security.Cryptography/test/Security.Cryptography/SipHashTests.AlgorithmName.cs
@@ -12,7 +12,7 @@ public abstract partial class SipHashTests<TTest, TAlgorithm>
     /// Verifies that <see cref="SipHash.AlgorithmName" />, when UsingVariant, returns the expected value.
     /// </summary>
     [TestMethod]
-    [DynamicData(nameof(HashAlgorithmVariants))]
+    [DynamicData(nameof(HashAlgorithmVariants), DynamicDataDisplayName = nameof(HashAlgorithmVariantDisplayName.GetDisplayName), DynamicDataDisplayNameDeclaringType = typeof(HashAlgorithmVariantDisplayName))]
     public void AlgorithmName_WhenUsingVariant_ShouldReturnCorrectlyFormattedString(SipHashVariant variant)
     {
         using TAlgorithm algorithm = CreateAlgorithm(variant);
